♻️ refactor(http): add utils
